#101 complete overhaul of accounts, signup flows and switching between stores

+++ app/controllers/accounts_controller.rb
@@ -8,20 +8,33 @@ class AccountsController < ApplicationController
 
   def join
     authorize @account
-    # Create an AccountUser if not exists
-    AccountUser.unscoped.where(account_id: @account.id, user_id: Current.user.id).first_or_create!(user_role: :store_manager)
+    ActsAsTenant.without_tenant do
+      AccountUser.unscoped.where(account: @account, user: Current.user).first_or_create!(user_role: :store_manager)
+    end
+  rescue ActiveRecord::RecordNotUnique
+    # Already exists, ignore
+  ensure
     session[:managed_account_id] = @account.id
-    # Force reload of current account/tenant for the current request context if needed,
-    # but here we are redirecting so session will be used in next request.
     redirect_to dashboard_path, notice: "Joined #{@account.name} as a support team member."
   end
 
   def leave
     authorize @account
-    # Remove the AccountUser
-    AccountUser.unscoped.where(account: @account, user: Current.user).delete_all
-    session.delete(:managed_account_id)
-    redirect_to accounts_path, notice: "Left support session for #{@account.name}."
+    # Only remove AccountUser if it was a support session (not the owner)
+    if Current.user.admin? && @account.owner_id != Current.user.id
+      AccountUser.unscoped.where(account: @account, user: Current.user).delete_all
+    end
+
+    # Use "none" to indicate explicitly no account selected, to prevent ApplicationController fallback
+    session[:managed_account_id] = "none"
+
+    notice = if Current.user.admin?
+      "Left support session for #{@account.name}."
+    else
+      "Stopped managing #{@account.name}."
+    end
+
+    redirect_to root_path, notice: notice
   end
 
   def audits
@@ -32,17 +45,20 @@ class AccountsController < ApplicationController
   # GET /accounts/1 or /accounts/1.json
   def show
     authorize @account
+    @account_users = @account.account_users.includes(:user).order("users.name")
   end
 
   # GET /accounts/new
   def new
     @account = Account.new(owner_id: Current.user.id)
     authorize @account
+    @tab = params[:tab] || "general"
   end
 
   # GET /accounts/1/edit
   def edit
     authorize @account
+    @tab = params[:tab] || "general"
   end
 
   # POST /accounts or /accounts.json
@@ -53,22 +69,12 @@ class AccountsController < ApplicationController
 
     respond_to do |format|
       if @account.save
-        # Create AccountUser for the owner
-        if @account.owner_id
-          AccountUser.unscoped.where(account_id: @account.id, user_id: @account.owner_id).first_or_create(user_role: :store_manager)
-        end
-
-        # Also make the creator an admin if different from owner
-        if Current.user&.id && Current.user.id != @account.owner_id.to_i
-          AccountUser.unscoped.where(account_id: @account.id, user_id: Current.user.id).first_or_create(user_role: :store_manager)
-        end
-
-        session[:managed_account_id] ||= @account.id
+        session[:managed_account_id] = @account.id
 
         if Current.user&.admin?
-          format.html { redirect_to @account, notice: "Account was successfully created." }
+          format.html { redirect_to params[:return_to] || @account, notice: "Account was successfully created." }
         else
-          format.html { redirect_to dashboard_path, notice: "Account was successfully created." }
+          format.html { redirect_to params[:return_to] || dashboard_path, notice: "Account was successfully created." }
         end
         format.json { render :show, status: :created, location: @account }
       else
@@ -111,6 +117,6 @@ class AccountsController < ApplicationController
 
     # Only allow a list of trusted parameters through.
     def account_params
-      params.require(:account).permit(:name, :address, :owner_id)
+      params.require(:account).permit(:name, :address, :owner_id, :header_image)
     end
 end
